# Optimize Bar component rendering when activeBar is falsy